% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga-descargar.R
\name{siga_datos}
\alias{siga_datos}
\alias{siga_metadatos}
\alias{siga_descargar}
\title{Descarga archivos del SIGA}
\usage{
siga_datos(ids, dir = tempdir(), forzar = FALSE)

siga_metadatos(ids, dir = tempdir(), forzar = FALSE)

siga_descargar(ids, dir = tempdir(), forzar = TRUE)
}
\arguments{
\item{ids}{Vector de ids con estaciones o un data.frame con una
columna llamada "id" con las ids de las estaciones (por ejemplo, el
devuelto por \code{\link[=siga_estaciones]{siga_estaciones()}}).}

\item{dir}{Directorio donde descargar los datos.}

\item{forzar}{Forzar la descarga de archivos ya descargados.}
}
\value{
data.frames
}
\description{
Descarga archivos del SIGA
}
