% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga-descargar.R
\name{siga_estaciones}
\alias{siga_estaciones}
\title{Estaciones disponibles en el SIGA}
\usage{
siga_estaciones(
  archivo = file.path(tempdir(), "siga_metadatos.csv"),
  forzar = FALSE
)
}
\arguments{
\item{archivo}{Caracter, indica la ruta al archivo donde guardar치 la lista de estaciones.
Por defecto se guarda en un directorio temporal.}

\item{forzar}{L칩gico, si es TRUE, fuerza la descarga del archivo de metadatos ya descargado.}
}
\value{
Un data.frame con metadatos de las estaciones.
}
\description{
La funci칩n descarga los metadatos de las estaciones meteorol칩gicas disponibles en el SIGA.
}
\examples{
estaciones <- siga_estaciones()
head(estaciones)

}
