% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga-descargar.R
\name{siga_datos}
\alias{siga_datos}
\alias{siga_metadatos}
\alias{siga_descargar}
\title{Descarga archivos del SIGA}
\usage{
siga_datos(ids, dir = tempdir(), forzar = FALSE)

siga_metadatos(ids, dir = tempdir(), forzar = FALSE)

siga_descargar(ids, dir = tempdir(), forzar = TRUE)
}
\arguments{
\item{ids}{Vector de ids con estaciones o un data.frame con una
columna llamada "id" con las ids de las estaciones (por ejemplo, el
data.frame devuelto por \code{\link[=siga_estaciones]{siga_estaciones()}}).}

\item{dir}{Caracter, directorio donde descargar los datos. Por defecto
se descargan los datos en un directorio temporal.}

\item{forzar}{Forzar la descarga de archivos ya descargados.}
}
\value{
data.frames
}
\description{
Esta familia de funciones permite leer y descargar archivos con datos
meteorológicos y sus metadatos desde SIGA.
}
\examples{
estaciones <- siga_estaciones()
ids <- estaciones[1:2, ]
siga_datos(ids)

# Metadatos
siga_metadatos(ids)

# Descarga directa
siga_descargar(ids)

}
