% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga-descargar.R
\name{siga_estaciones}
\alias{siga_estaciones}
\title{Estaciones disponibles en SIGA}
\usage{
siga_estaciones(
  archivo = file.path(tempdir(), "siga_metadatos.csv"),
  forzar = FALSE
)
}
\arguments{
\item{archivo}{Caracter con la ruta al archivo donde guardar la lista de estaciones.}

\item{forzar}{Forzar la descarga de archivos ya descargados.}
}
\value{
Un data.frame
}
\description{
Estaciones disponibles en SIGA
}
